refactor submodule to file
